# docs/assets/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/api-key": {
      "get": {
        "operationId": "ApiKeyController_getApiKey",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns the user's API key if it exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyInfoWrapperDto"
                }
              }
            }
          },
          "401": {
            "description": "User is not authenticated"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          }
        ],
        "summary": "Get the current user's API key information",
        "tags": [
          "API Keys"
        ]
      },
      "post": {
        "operationId": "ApiKeyController_generateApiKey",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Returns the newly generated API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedApiKeyWrapperDto"
                }
              }
            }
          },
          "401": {
            "description": "User is not authenticated"
          },
          "409": {
            "description": "User already has an API key. Delete it first or use regenerate."
          }
        },
        "security": [
          {
            "keycloak-sso": []
          }
        ],
        "summary": "Generate a new API key for the current user",
        "tags": [
          "API Keys"
        ]
      },
      "delete": {
        "operationId": "ApiKeyController_deleteApiKey",
        "parameters": [],
        "responses": {
          "204": {
            "description": "API key deleted successfully"
          },
          "401": {
            "description": "User is not authenticated"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          }
        ],
        "summary": "Delete the current user's API key",
        "tags": [
          "API Keys"
        ]
      }
    },
    "/api/api-key/regenerate": {
      "post": {
        "operationId": "ApiKeyController_regenerateApiKey",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns the newly generated API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedApiKeyWrapperDto"
                }
              }
            }
          },
          "401": {
            "description": "User is not authenticated"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          }
        ],
        "summary": "Regenerate the current user's API key",
        "tags": [
          "API Keys"
        ]
      }
    },
...
  },
  "info": {
    "title": "Document Intelligence Platform API",
    "description": "Comprehensive REST API for document processing, OCR, workflow orchestration, model training, and human-in-the-loop review",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "keycloak-sso": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "ApiKeyInfoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "keyPrefix": {
            "type": "string"
          },
          "userEmail": {
            "type": "string"
          },
          "roles": {
            "description": "Roles inherited from the creating user at key generation time",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "lastUsed": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "keyPrefix",
          "userEmail",
          "roles",
          "createdAt",
          "lastUsed"
        ]
      },
      "ApiKeyInfoWrapperDto": {
        "type": "object",
        "properties": {
          "apiKey": {
            "$ref": "#/components/schemas/ApiKeyInfoDto"
          }
        },
        "required": [
          "apiKey"
        ]
      },
      "GeneratedApiKeyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "keyPrefix": {
            "type": "string"
          },
          "userEmail": {
            "type": "string"
          },
          "roles": {
            "description": "Roles inherited from the creating user at key generation time",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "lastUsed": {
            "format": "date-time",
            "type": "string"
          },
          "key": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "keyPrefix",
          "userEmail",
          "roles",
          "createdAt",
          "lastUsed",
          "key"
        ]
      },
      "GeneratedApiKeyWrapperDto": {
        "type": "object",
        "properties": {
          "apiKey": {
            "$ref": "#/components/schemas/GeneratedApiKeyDto"
          }
        },
        "required": [
          "apiKey"
        ]
      },
...
    }
  }
}